requirements.txt
Image
Pillow
docopt
numpy
colormath
